HW1 - From Q-Learning to Deep Q-learning/requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
bleach==4.0.0
blinker==1.4
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
coverage==5.5
cryptography==3.4.8
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fonttools==4.25.0
gast==0.3.3
glfw==2.4.0
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.36.1
gym==0.21.0
h5py==2.10.0
idna==3.2
imageio==2.10.4
importlib-metadata==4.8.1
importlib-resources==5.4.0
ipykernel==6.4.1
ipython==7.29.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.2
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lockfile==0.12.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.1.0
munkres==1.1.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numexpr==2.7.3
numpy==1.18
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyglet==1.5.21
Pygments==2.10.0
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==3.0.4
pyreadline==2.1
pyrsistent==0.18.0
PySocks==1.7.1
python-box==5.4.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==228
pywinpty==0.5.7
PyYAML==6.0
pyzmq==22.2.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scipy==1.4.1
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==58.0.4
sip>=4.19.13
six==1.16.0
tensorboard==2.4.0
tensorboard-plugin-wit>=1.6.0
tensorboardX==2.4
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wheel==0.37.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.3
zipp==3.6.0
